make filter of subtrack consistent with its state
