+++ __tests__/shared/components/challenge-listing/Filters/FiltersPanel.jsx
@@ -25,7 +25,7 @@ const mockDatas = [{
   setFilterState,
   setSearchText,
   validKeywords: ['key', 'word'],
-  validSubtracks: [{name: 'sub', subTrack: 'sub'}, {name: 'track', subTrack: 'track'}],
+  validSubtracks: [{ name: 'sub', subTrack: 'sub' }, { name: 'track', subTrack: 'track' }],
   onClose,
 }, {
   communityFilters: [
